+++ diesel_cli/tests/print_schema.rs
@@ -116,6 +116,20 @@ fn print_schema_custom_types() {
 }
 
 #[test]
+fn print_schema_with_unmappable_names() {
+    test_print_schema("print_schema_with_unmappable_names", vec!["--with-docs"]);
+}
+
+#[test]
+#[cfg(feature = "postgres")]
+fn print_schema_with_unmappable_names_and_schema_name() {
+    test_print_schema(
+        "print_schema_with_unmappable_names_and_schema_name",
+        vec!["--with-docs", "--schema", "custom_schema"],
+    )
+}
+
+#[test]
 fn schema_file_is_relative_to_project_root() {
     let p = project("schema_file_is_relative_to_project_root")
         .folder("foo")